src file uploaded for carla_ws/
The src folder includes a bash file to launch CarlaUE4 from prior install. setup.py file with imported packages and entry points for scripts I want to launch which included spawning of a vehicle and sensors. The script will make ros2 topics based on the sensors attached to vehicle along with their transform.